Get all todos based on user id

+++ TodoApp/routers/todos.py
@@ -1,14 +1,14 @@
 from typing import Annotated
 from pydantic import BaseModel, Field
 from sqlalchemy.orm import Session
-from fastapi import Depends, HTTPException, Path,APIRouter
+from fastapi import Depends, HTTPException, Path, APIRouter
 from starlette import status
 from models import Todos
 from database import SessionLocal
 from .auth import get_current_user
 
-
 router = APIRouter()
+
 
 def get_db():
     db = SessionLocal()
@@ -18,46 +18,57 @@ def get_db():
         db.close()
 
 
-db_dependency = Annotated[Session,Depends(get_db)]
-user_dependency = Annotated[dict,Depends(get_current_user)]
+db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
-    title : str = Field(min_length=3)
-    description : str = Field(min_length=3,max_length=100)
-    priority : int = Field(gt=0,lt=6)
-    completed : bool
-
-@router.get("/",status_code=status.HTTP_200_OK)
-async def read_all(user:user_dependency,db: db_dependency):
-    return db.query(Todos).filter(Todos.owner_id==user.get('id')).all()
+    title: str = Field(min_length=3)
+    description: str = Field(min_length=3, max_length=100)
+    priority: int = Field(gt=0, lt=6)
+    completed: bool
 
 
-@router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
-async def read_todo(db:db_dependency,todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.get("/", status_code=status.HTTP_200_OK)
+async def read_all(user: user_dependency, db: db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
+
+@router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
+async def read_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
-    raise HTTPException(status.HTTP_404_NOT_FOUND)
+    raise HTTPException(status.HTTP_404_NOT_FOUND, detail="Todo not found")
 
-@router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(user:user_dependency,
-                      db:db_dependency,
-                      todo_request:TodoRequest):
+
+@router.post("/todo", status_code=status.HTTP_201_CREATED)
+async def create_todo(user: user_dependency,
+                      db: db_dependency,
+                      todo_request: TodoRequest):
     if user is None:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,detail="Authentication Failed")
-    todo_model = Todos(**todo_request.dict(),owner_id=user.get('id'))
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = Todos(**todo_request.dict(), owner_id=user.get('id'))
     db.add(todo_model)
     db.commit()
 
 
-@router.put("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db:db_dependency,
-                        todo_request:TodoRequest,
-                        todo_id:int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def update_todo(user: user_dependency, db: db_dependency,
+                      todo_request: TodoRequest,
+                      todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id) \
+        .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
+
     todo_model.title = todo_request.title
     todo_model.description = todo_request.description
     todo_model.priority = todo_request.priority
@@ -67,11 +78,13 @@ async def update_todo(db:db_dependency,
     db.commit()
 
 
-@router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db:db_dependency,todo_id:int=Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id). \
+        filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND)
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Todo not found')
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
     db.commit()
-
